+++ gem/forest-client.gemspec
@@ -1,0 +1,12 @@
+Gem::Specification.new do |s|
+    s.name = 'forest-client'
+    s.version = '0.0.8'
+    s.date = '2013-11-15'
+    s.summary = 'Forest Printer Management Query Agent'
+    s.description = 'The LAN printer query agent for the Forest printer management system.'
+    s.authors = ['Ben Burwell', 'Andrew Trautman', 'Macauley Breault', 'Mike Borowski']
+    s.email = 'printer.fungineers@gmail.com'
+    s.files = ['lib/forest-client.rb', 'lib/forest-client/printer.rb']
+    s.homepage = 'https://github.com/printerSystemCSI210'
+    s.license = 'Unknown'
+end
