--- lib/snmp.rb
@@ -1,11 +1,0 @@
-#
-# Copyright (c) 2004 David R. Halliday
-# All rights reserved.
-#
-# This SNMP library is free software.  Redistribution is permitted under the
-# same terms and conditions as the standard Ruby distribution.  See the
-# COPYING file in the Ruby distribution for details.
-#
-
-require 'snmp/manager'
-require 'snmp/version'
