--- lib/snmp/version.rb
@@ -1,3 +1,0 @@
-module SNMP
-  VERSION = "1.1.1"
-end
