+++ test_gem.rb
@@ -1,0 +1,4 @@
+require 'forest-client'
+
+qa = ForestClient::QueryAgent.new('527add43c1781f0200000001')
+qa.run
